+++ Python/palindrome.py
